fixed #107
