+++ src/components/AudienceItem.tsx
@@ -1,18 +1,35 @@
 import React from "react";
+import { connect } from "react-redux";
+import { ThunkDispatch } from "redux-thunk";
+import { AnyAction } from "redux";
 
+import { TStore } from "../store";
+import { selectPeer } from "../actions";
 import { Member } from "../types";
 
 import "./AudienceItem.css";
 
 interface IProps {
   audience: Member;
-  isMuted: boolean;
+  isMine: boolean;
   isSelected: boolean;
+  selectPeer: any;
 }
 
 class AudienceItem extends React.PureComponent<IProps> {
   // As video.playsInline is not defined in HTMLVideoElement, add "any" as well.
   private _videoRef = React.createRef<HTMLVideoElement | any>();
+
+  constructor(props: IProps) {
+    super(props);
+
+    this._onClick = this._onClick.bind(this);
+  }
+
+  _onClick() {
+    const { audience, selectPeer } = this.props;
+    selectPeer(audience.peerId);
+  }
 
   _updateVideo() {
     const video = this._videoRef.current;
@@ -35,17 +52,25 @@ class AudienceItem extends React.PureComponent<IProps> {
   }
 
   render() {
-    const { isSelected, isMuted, audience } = this.props;
+    const { isSelected, isMine, audience } = this.props;
     return (
       <li
         className={
-          "audience-item " + (isSelected ? "audience-item--selected" : "")
+          "audience-item" +
+          (isSelected ? " audience-item--selected" : "") +
+          (isMine ? " audience-item--mine" : "")
         }
+        onClick={this._onClick}
       >
-        <img src={audience.dataURL} className="audience-item__icon"></img>
+        <img
+          src={audience.dataURL}
+          className={
+            "audience-item__icon" + (isMine ? " audience-item__icon--mine" : "")
+          }
+        ></img>
         <video
           className="audience-item__video"
-          muted={isMuted}
+          muted={isMine}
           ref={this._videoRef}
         />
       </li>
@@ -53,4 +78,12 @@ class AudienceItem extends React.PureComponent<IProps> {
   }
 }
 
-export default AudienceItem;
+const mapDispatchToProps = (
+  dispatch: ThunkDispatch<TStore, void, AnyAction>
+) => ({
+  selectPeer: (peerId: string) => {
+    dispatch(selectPeer(peerId));
+  },
+});
+
+export default connect(null, mapDispatchToProps)(AudienceItem);
