--- src/components/PresenterSelecting.css
@@ -1,17 +1,0 @@
-.presenter-selecting {
-  width: 14vw;
-  height: 14vw;
-  max-width: 120px;
-  max-height: 120px;
-  display: inline-block;
-  pointer-events: auto;
-  background-color: transparent;
-  background-size: cover;
-  border: none;
-  outline: none;
-}
-
-.presenter-selecting__image {
-  width: 100%;
-  height: 100%;
-}
