+++ src/components/Room.css
@@ -3,11 +3,3 @@
   width: 100%;
   height: 100%;
 }
-
-.room__viewtoollist {
-  display: flex;
-  position: absolute;
-  bottom: 0;
-  width: 28vw;
-  max-width: 240px;
-}
